+++ src/hibiki.ts
@@ -16,7 +16,7 @@ import {HibikiRootNode, CORE_LIBRARY} from "./nodes";
 import {deepTextContent, evalDeepTextContent, isObject, bindLibContext, callHook, parseTextData} from "./utils";
 import * as HibikiUtilsModule from "./utils";
 import merge from "lodash/merge";
-import type {HibikiConfig, Hibiki, HibikiExtState, ReactClass, LibraryType, HibikiGlobalConfig} from "./types";
+import type {HibikiConfig, Hibiki, HibikiExtState, ReactClass, LibraryType, HibikiGlobalConfig, HtmlParserOpts} from "./types";
 import type {HibikiNode} from "./html-parser";
 import {LocalModule, HttpModule, LibModule} from "./modules";
 import {HibikiModule} from "./hibiki-module";
@@ -95,6 +95,15 @@ function readHibikiConfigFromOuterHtml(htmlElem : string | HTMLElement) : Hibiki
     if (htmlElem.hasAttribute("name")) {
         rtn.stateName = htmlElem.getAttribute("name");
     }
+    if (htmlElem.hasAttribute("textdelimiters")) {
+        let delimType = htmlElem.getAttribute("textdelimiters");
+        if (delimType === "none" || delimType === "default" || delimType === "alt") {
+            rtn.textDelimiters = delimType;
+        }
+        else {
+            console.log(sprintf("Invalid 'textdelimiters' attribute on hibiki tag, allowed values are 'none', 'default', or 'alt', got='%s'", delimType));
+        }
+    }
     return rtn;
 }
 
@@ -102,21 +111,26 @@ function readHibikiConfigFromOuterHtml(htmlElem : string | HTMLElement) : Hibiki
 let createState = function createState(config : HibikiConfig, html : string | HTMLElement, initialData : any) : HibikiExtState {
     let state = new HibikiState();
     state.ComponentLibrary.addLibrary(CORE_LIBRARY);
+
+    let configFromOuterHtml = readHibikiConfigFromOuterHtml(html);
+    config = merge({}, configFromOuterHtml, config);
     
-    config = config || {};
     initialData = initialData || {};
     let htmlObj : HibikiNode = null;
+    let parserOpts : HtmlParserOpts = {};
+    if (config.textDelimiters != null) {
+        parserOpts.textDelimiters = config.textDelimiters;
+    }
     if (config.htmlSrc != null) {
-        htmlObj = parseHtml(config.htmlSrc);
+        htmlObj = parseHtml(config.htmlSrc, "config.htmlSrc", parserOpts);
         bindLibContext(htmlObj, "main");
     }
     else if (html != null) {
-        htmlObj = parseHtml(html);
+        let sourceName = (typeof(html) === "string" ? "html-string" : sprintf("<%s>", html.tagName.toLowerCase()));
+        htmlObj = parseHtml(html, sourceName, parserOpts);
         bindLibContext(htmlObj, "main");
     }
     state.setHtml(htmlObj);
-    let configFromOuterHtml = readHibikiConfigFromOuterHtml(html);
-    config = merge({}, config, configFromOuterHtml);
     let hibikiOpts = readHibikiOptsFromHtml(htmlObj);
     if (!config.noConfigMergeFromHtml) {
         config = merge({}, (hibikiOpts.config ?? {}), config);