+++ changelog.md
@@ -3,6 +3,7 @@
 ## v0.3.4
 
 * Allow parsing of YAML data if js-yaml UMD package is manually included (window.jsyaml is set)
+* Can set textDelimiters (on config and on hibiki tag as an attribute) to "none", "default", or "alt".  Selects between no inline h-text elements, {{ ... }}, and {| ... |} respectively.  The "alt" option can be used with Django templates.
 * do not fire welcome message until DOMContentLoaded event
 * new function js:moveitem(arr, from-index, to-index)
 * expand foreach inside of \@children
@@ -11,6 +12,7 @@
 * export HibikiUtils module for libraries
 * new element geometry tracking: geo.bindpath (client/offset height/width, and scrolltop)
 * (beta) new scroll tracking: scrolltop.bindpath, and scroll.handler (debounced)
+* remove ability to set window.HibikiParserOpts
 * bugfix: reload HibikiGlobalConfig after DOMContentLoaded event
 * bugfix: Hibiki.States was not holding the correct HibikiExtState object